Add helper function to create query client for tests.
