+++ coral/src/app/features/browse-topics/hooks/environment/useGetEnvironments.test.tsx
@@ -1,21 +1,15 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import { ReactElement } from "react";
 import { server } from "src/services/api-mocks/server";
 import { useGetEnvironments } from "src/app/features/browse-topics/hooks/environment/useGetEnvironments";
 import { mockGetEnvironments } from "src/domain/environment";
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: false,
-      cacheTime: Infinity,
-    },
-  },
-});
+import { getQueryClientForTests } from "src/services/test-utils/query-client-tests";
 
 const wrapper = ({ children }: { children: ReactElement }) => (
-  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  <QueryClientProvider client={getQueryClientForTests()}>
+    {children}
+  </QueryClientProvider>
 );
 
 describe("useGetEnvironments", () => {
