+++ coral/src/services/test-utils/query-client-tests.ts
@@ -1,0 +1,14 @@
+import { QueryClient } from "@tanstack/react-query";
+
+function getQueryClientForTests() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        retry: false,
+        cacheTime: Infinity,
+      },
+    },
+  });
+}
+
+export { getQueryClientForTests };
